% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konwersja_temp.R
\name{konwersja_temp}
\alias{konwersja_temp}
\title{Konwersja temperatury}
\usage{
konwersja_temp(temperatura_f)
}
\arguments{
\item{temperatura_f}{temperatura w F}
}
\value{
wektor o tej samej dlugosci jak temperatura
}
\description{
Konwersja temperatury
}
